Creature Data
Currently iterating upon the Creature data structure until it proves suitable for a truly modular system. The idea is that the ruleset file will contain a set of parameters that inform the program on how to generate any type of game asset. Later, a web interface will be introduced to browse the content similar to a virtual novel, but with a HUD for the things you'd generally keep track of with paper and cardboard assets in a tabletop rpg, and a hands-off approach to enforcing the mechanics like the Fantasy Flight games Descent and Mansions of Madness.